backend/requirements.txt
fastapi
uvicorn
httpx
openai
groq
supabase
python-dotenv
starlette